+++ samba_on.sh
@@ -24,33 +24,33 @@
 
 
 
-if [ "$(uname -n)" != "MiSTer" ]
+if [ -f "/media/fat/MiSTer" ]; 
 then
+	mount | grep "on / .*[(,]ro[,$]" -q && RO_ROOT="true"
+	[ "$RO_ROOT" == "true" ] && mount / -o remount,rw
+	mv /etc/init.d/_S91smb /etc/init.d/S91smb > /dev/null 2>&1
+	sync
+	[ "$RO_ROOT" == "true" ] && mount / -o remount,ro
+	if [ -f /media/fat/linux/iptables.up.rules ]
+	then
+		sed -e '/--dport 137 /s/^#//g' -i /media/fat/linux/iptables.up.rules
+		sed -e '/--dport 138 /s/^#//g' -i /media/fat/linux/iptables.up.rules
+		sed -e '/--dport 139 /s/^#//g' -i /media/fat/linux/iptables.up.rules
+		sed -e '/--dport 445 /s/^#//g' -i /media/fat/linux/iptables.up.rules
+	fi
+	sync
+	if [ -f /etc/network/if-pre-up.d/iptables ]
+	then
+		/etc/network/if-pre-up.d/iptables
+	fi
+	/etc/init.d/S91smb start
+
+	echo "Samba is on and"
+	echo "active at startup."
+	echo "Done!"
+	exit 0
+else
 	echo "This script should be run"
 	echo "on a MiSTer system."
 	#exit 1
 fi
-
-mount | grep "on / .*[(,]ro[,$]" -q && RO_ROOT="true"
-[ "$RO_ROOT" == "true" ] && mount / -o remount,rw
-mv /etc/init.d/_S91smb /etc/init.d/S91smb > /dev/null 2>&1
-sync
-[ "$RO_ROOT" == "true" ] && mount / -o remount,ro
-if [ -f /media/fat/linux/iptables.up.rules ]
-then
-	sed -e '/--dport 137 /s/^#//g' -i /media/fat/linux/iptables.up.rules
-	sed -e '/--dport 138 /s/^#//g' -i /media/fat/linux/iptables.up.rules
-	sed -e '/--dport 139 /s/^#//g' -i /media/fat/linux/iptables.up.rules
-	sed -e '/--dport 445 /s/^#//g' -i /media/fat/linux/iptables.up.rules
-fi
-sync
-if [ -f /etc/network/if-pre-up.d/iptables ]
-then
-	/etc/network/if-pre-up.d/iptables
-fi
-/etc/init.d/S91smb start
-
-echo "Samba is on and"
-echo "active at startup."
-echo "Done!"
-exit 0
